refactor test folder structure
