+++ .github/workflows/main_tests.yml
@@ -35,7 +35,7 @@ jobs:
       - name: Test conda environment
         run: |
           conda activate soprano-dev
-          pytest -s tests/test_configuration
+          pytest -s tests/test_installation.py
       - name: Test units
         run: |
           conda activate soprano-dev
